+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: rpohlen <rpohlen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/11 17:34:18 by rpohlen           #+#    #+#             */
-/*   Updated: 2021/11/24 19:01:59 by rpohlen          ###   ########.fr       */
+/*   Updated: 2021/11/25 16:01:52 by rpohlen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ void	ft_bzero(void *s, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
-int		ft_strlcat(char *dest, const char *src, size_t n);
+size_t	ft_strlcat(char *dest, const char *src, size_t n);
 int		ft_tolower(int c);
 int		ft_toupper(int c);
 char	*ft_strchr(const char *str, int c);
